+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
   - Do not throw after handling a port conflict error when creating a proxy server.
 
+### Fixed
+
+  - Fix syntax error display regression.
+
 ## [9.1.0] - 2019-04-30
 
 ### Added